EHN: defining safe log as activation inside module
